app/requirements.txt
faiss_cpu==1.8.0.post1
streamlit==1.43.2
langchain==0.3.21
langchain_openai==0.3.11
langgraph==0.3.21
nltk==3.9.1
numpy
openai==1.69.0
pandas==2.2.3
psycopg2_binary==2.9.9
pymorphy3==2.0.2
python-dotenv==1.1.0
rank_bm25==0.2.2
semantic_router==0.1.6
sentence_transformers==3.0.1